reorg(extrusions) split out extrudeRectangularPath2 and extrudeRectangularGeom2
