+++ src/operations/extrusions/extrudeRectangular.js
@@ -2,63 +2,8 @@ const flatten = require('../../utils/flatten')
 
 const { geom2, path2 } = require('../../geometry')
 
-const { expand } = require('../expansions')
-
-const area = require('../expansions/area')
-
-const extrudeLinear = require('./extrudeLinear')
-
-/*
- * Expand and extrude the given path.
- */
-const extrudeFromPath2 = (options, geometry) => {
-  const defaults = {
-    radius: 1,
-    segments: 0
-  }
-  let { radius, segments } = Object.assign({ }, defaults, options)
-
-  options.delta = radius
-  options.segments = segments
-
-  const points = path2.toPoints(geometry)
-  if (points.length === 0) throw new Error('the given geometry cannot be empty')
-
-  let newgeometry = expand(options, geometry)
-  return extrudeLinear(options, newgeometry)
-}
-
-/*
- * Expand and extrude the given geom2.
- */
-const extrudeFromGeom2 = (options, geometry) => {
-  const defaults = {
-    radius: 1,
-    segments: 0
-  }
-  let { radius, segments } = Object.assign({ }, defaults, options)
-
-  options.delta = radius
-  options.segments = segments
-
-  // convert the geometry to outlines
-  let outlines = geom2.toOutlines(geometry)
-  if (outlines.length === 0) throw new Error('the given geometry cannot be empty')
-
-  // expand the outlines
-  let newparts = outlines.map((outline) => {
-    if (area(outline) < 0) outline.reverse() // all outlines must wind counter clockwise
-    return expand(options, path2.fromPoints({ closed: true }, outline))
-  })
-
-  // create a composite geometry
-  let allsides = newparts.reduce((sides, part) => {
-    return sides.concat(geom2.toSides(part))
-  }, [])
-  let newgeometry = geom2.create(allsides)
-
-  return extrudeLinear(options, newgeometry)
-}
+const extrudeRectangularPath2 = require('./extrudeRectangularPath2')
+const extrudeRectangularGeom2 = require('./extrudeRectangularGeom2')
 
 /**
  * Extrude the given object(s) by following the outline(s) with a rectangle.
@@ -88,8 +33,8 @@ const extrudeRectangular = (options, ...objects) => {
   if (segments < 0) throw new Error('segments must be zero or more')
 
   const results = objects.map((object) => {
-    if (path2.isA(object)) return extrudeFromPath2(options, object)
-    if (geom2.isA(object)) return extrudeFromGeom2(options, object)
+    if (path2.isA(object)) return extrudeRectangularPath2(options, object)
+    if (geom2.isA(object)) return extrudeRectangularGeom2(options, object)
     // if (geom3.isA(object)) return geom3.transform(matrix, object)
     return object
   })
